campaign-backend/requirements.txt
flask
flask-sqlalchemy
psycopg2
flask-migrate
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
statsmodels>=0.12.0
requests
rollbar
python-dotenv
gunicorn
flask_cors
